requirements.txt
azure-eventhub==5.8.0
azure-storage-blob==12.13.1
pyspark==3.2.0
numpy==1.21.5
pandas==1.4.2
scikit-learn==1.0.2
matplotlib==3.5.1
python-dotenv==0.19.2
jupyter==1.0.0 